refactor[#217] : api get 로직 features -> entities 로 이동
