+++ src/entities/profile/constants/endpoint.ts
